Move ped-anova tests to test_evaluator.py
