+++ tests/importance_tests/pedanova_tests/test_evaluator.py
@@ -3,11 +3,14 @@ from __future__ import annotations
 from collections.abc import Callable
 from copy import deepcopy
 
+import numpy as np
 import pytest
 
 import optuna
+from optuna.importance import PedAnovaImportanceEvaluator
 from optuna.importance._ped_anova.evaluator import _QuantileFilter
 from optuna.trial import FrozenTrial
+from tests.importance_tests.test_importance_evaluators import get_study
 
 
 _VALUES = list([[float(i)] for i in range(10)])[::-1]
@@ -49,3 +52,51 @@ def test_filter(
     indices = [t.user_attrs["index"] for t in _filter.filter(trials)]
     assert len(indices) == len(filtered_indices)
     assert all(i == j for i, j in zip(indices, filtered_indices))
+
+
+def test_error_in_ped_anova() -> None:
+    with pytest.raises(RuntimeError):
+        evaluator = PedAnovaImportanceEvaluator()
+        study = get_study(seed=0, n_trials=5, is_multi_obj=True)
+        evaluator.evaluate(study)
+
+
+def test_n_trials_equal_to_min_n_top_trials() -> None:
+    evaluator = PedAnovaImportanceEvaluator()
+    study = get_study(seed=0, n_trials=evaluator._min_n_top_trials, is_multi_obj=False)
+    param_importance = list(evaluator.evaluate(study).values())
+    n_params = len(param_importance)
+    assert np.allclose(param_importance, np.full(n_params, 1.0 / n_params))
+
+
+def test_baseline_quantile_is_1() -> None:
+    study = get_study(seed=0, n_trials=100, is_multi_obj=False)
+    # baseline_quantile=1.0 enforces top_trials == all_trials identical.
+    evaluator = PedAnovaImportanceEvaluator(baseline_quantile=1.0)
+    param_importance = list(evaluator.evaluate(study).values())
+    n_params = len(param_importance)
+    # When top_trials == all_trials, all the importances become identical.
+    assert np.allclose(param_importance, np.full(n_params, 1.0 / n_params))
+
+
+def test_direction() -> None:
+    study_minimize = get_study(seed=0, n_trials=20, is_multi_obj=False)
+    study_maximize = optuna.create_study(direction="maximize")
+    study_maximize.add_trials(study_minimize.trials)
+
+    evaluator = PedAnovaImportanceEvaluator()
+    assert evaluator.evaluate(study_minimize) != evaluator.evaluate(study_maximize)
+
+
+def test_baseline_quantile() -> None:
+    study = get_study(seed=0, n_trials=20, is_multi_obj=False)
+    default_evaluator = PedAnovaImportanceEvaluator(baseline_quantile=0.1)
+    evaluator = PedAnovaImportanceEvaluator(baseline_quantile=0.3)
+    assert evaluator.evaluate(study) != default_evaluator.evaluate(study)
+
+
+def test_evaluate_on_local() -> None:
+    study = get_study(seed=0, n_trials=20, is_multi_obj=False)
+    default_evaluator = PedAnovaImportanceEvaluator(evaluate_on_local=True)
+    global_evaluator = PedAnovaImportanceEvaluator(evaluate_on_local=False)
+    assert global_evaluator.evaluate(study) != default_evaluator.evaluate(study)
