+++ tests/importance_tests/test_importance_evaluators.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import numpy as np
 import pytest
 
 from optuna import create_study
@@ -73,45 +72,6 @@ def _test_max_depth_of_tree_based_evaluator(
     assert param_importance != param_importance_different_max_depth
 
 
-def test_small_n_trials_for_ped_anova() -> None:
-    study = get_study(seed=0, n_trials=2, is_multi_obj=False)
-    evaluator = PedAnovaImportanceEvaluator()
-    param_importance = list(evaluator.evaluate(study).values())
-    n_params = len(param_importance)
-    assert np.allclose(param_importance, np.full(n_params, 1.0 / n_params))
-
-
-def test_small_n_trials() -> None:
-    study = get_study(seed=0, n_trials=100, is_multi_obj=False)
-    evaluator = PedAnovaImportanceEvaluator(baseline_quantile=1.0)
-    param_importance = list(evaluator.evaluate(study).values())
-    n_params = len(param_importance)
-    assert np.allclose(param_importance, np.full(n_params, 1.0 / n_params))
-
-
-def test_direction_of_ped_anova() -> None:
-    study_minimize = get_study(seed=0, n_trials=20, is_multi_obj=False)
-    study_maximize = create_study(direction="maximize")
-    study_maximize.add_trials(study_minimize.trials)
-
-    evaluator = PedAnovaImportanceEvaluator()
-    assert evaluator.evaluate(study_minimize) != evaluator.evaluate(study_maximize)
-
-
-def test_baseline_quantile_of_ped_anova() -> None:
-    study = get_study(seed=0, n_trials=20, is_multi_obj=False)
-    default_evaluator = PedAnovaImportanceEvaluator()
-    evaluator = PedAnovaImportanceEvaluator(baseline_quantile=0.3)
-    assert evaluator.evaluate(study) != default_evaluator.evaluate(study)
-
-
-def test_evaluate_on_local_of_ped_anova() -> None:
-    study = get_study(seed=0, n_trials=20, is_multi_obj=False)
-    default_evaluator = PedAnovaImportanceEvaluator()
-    global_evaluator = PedAnovaImportanceEvaluator(evaluate_on_local=False)
-    assert global_evaluator.evaluate(study) != default_evaluator.evaluate(study)
-
-
 def _test_evaluator_with_infinite(
     evaluator_cls: type[BaseImportanceEvaluator], inf_value: float, target_idx: int | None = None
 ) -> None:
@@ -167,13 +127,6 @@ def _test_evaluator_with_only_non_single_dists(
 
     param_importance = evaluator.evaluate(study)
     assert param_importance == {}
-
-
-def test_error_in_ped_anova() -> None:
-    with pytest.raises(RuntimeError, match=r".*multi-objective optimization.*"):
-        evaluator = PedAnovaImportanceEvaluator()
-        study = get_study(seed=0, n_trials=5, is_multi_obj=True)
-        evaluator.evaluate(study)
 
 
 @pytest.mark.parametrize(
